Generates Page HTML Templates
Each page needs to be wrapped in a basic HTML template. More will be
added to this in the future, but for now a simple template can be used
for all pages.

I was bummed that I had to introduce some logic for building the data
structure passed to fast-xml-parser in the Element object hierarchy,
but when I attempted to move it to the Builder it left a lot of class
detection logic that probably would be less maintainable. So I'll have
to live with one of those design smells for now, and this seemed the
lesser of two evils.