Component Evaluation
This introduces the logic to parse component files. It brings the
logic all the way up to the build command, and passes the resulting
environment into page construction. It stops short of actually doing
anything with it when constructing the page, though, as that's another
project.

I spent some time thinking about introducing some notion of importing,
but ultimately concluded that it would just cause noise. The primary
benefit of that pattern is to allow a namespace in large projects, and
it seems to me like that can also be achieved by having several
smaller clusters of pages. Shared functionality could be achieved
through some form of modules, which I was planning on building
anyway.