Do Not Process Entities
I'm slowly shaping fast-xml-parser's options into what I need for this
project. The general principle here is that I want a set of options
in which XMLBuilder produces the same value that was passed into
XMLParser if no modifications occur. In this case, people will still
want to specify entities in their code, which this allows.